# PERF: significantly improve performance for nanmin/max for axis=None