--- a/bottleneck/src/reduce_template.c
+++ b/bottleneck/src/reduce_template.c
@@ -428,13 +428,14 @@ REDUCE_MAIN(NAME, 1)
 /* nanmin, nanmax -------------------------------------------------------- */
 
 /* repeat = {'NAME':      ['nanmin',         'nanmax'],
+             'INTCOMP':   ['<',              '>'],
              'COMPARE':   ['<=',             '>='],
              'BIG_FLOAT': ['BN_INFINITY',    '-BN_INFINITY'],
              'BIG_INT':   ['NPY_MAX_DTYPE0', 'NPY_MIN_DTYPE0']} */
 /* dtype = [['float64'], ['float32']] */
+BN_OPT_3
 REDUCE_ALL(NAME, DTYPE0) {
-    npy_DTYPE0 ai, extreme = BIG_FLOAT;
-    int allnan = 1;
+    npy_DTYPE0 extreme = BIG_FLOAT;
     INIT_ALL
     if (SIZE == 0) {
         VALUE_ERR("numpy.NAME raises on a.size==0 and axis=None; "
@@ -442,17 +443,67 @@ REDUCE_ALL(NAME, DTYPE0) {
         return NULL;
     }
     BN_BEGIN_ALLOW_THREADS
-    WHILE {
-        FOR {
-            ai = AI(DTYPE0);
-            if (ai COMPARE extreme) {
-                extreme = ai;
-                allnan = 0;
+    npy_bool allnan = 1;
+    if (REDUCE_CONTIGUOUS) {
+        const npy_intp LOOP_SIZE = 512 / sizeof(npy_DTYPE0);
+        // Allow vectorization by keeping an array of extremums
+        npy_DTYPE0 extremes[LOOP_SIZE];
+        for (npy_intp i=0; i < LOOP_SIZE; i++) {
+            extremes[i] = BIG_FLOAT;
+        }
+        const npy_intp loops = LENGTH / LOOP_SIZE;
+        const npy_intp residual = LENGTH % LOOP_SIZE;
+        WHILE {
+            const npy_DTYPE0* pa = PA(DTYPE0);
+            for (npy_intp i=0; i < loops; i++) {
+                if (allnan) {
+                    for (npy_intp j=0; j < LOOP_SIZE; j++) {
+                        const npy_DTYPE0 ai = pa[i * LOOP_SIZE + j];
+                        // We only need <=/>= until we identify our first non-nan value
+                        if (ai COMPARE extremes[j]) {
+                            allnan = 0;
+                            extremes[j] = ai;
+                        }
+                    }
+                } else {
+                    // Fully vectorize since the result cannot be NaN now
+                    for (npy_intp j=0; j < LOOP_SIZE; j++) {
+                        const npy_DTYPE0 ai = pa[i * LOOP_SIZE + j];
+                        extremes[j] = ai INTCOMP extremes[j] ? ai : extremes[j];
+                    }
+                }
+            }
+            for (npy_intp j=0; j < residual; j++) {
+                const npy_DTYPE0 ai = pa[loops * LOOP_SIZE + j];
+                if (ai COMPARE extremes[j]) {
+                    allnan = 0;
+                    extremes[j] = ai;
+                }
+            }
+            NEXT
+        }
+        // Reduce our array to find the global max/min
+        for (npy_intp j=0; j < LOOP_SIZE; j++) {
+            if (extremes[j] COMPARE extreme) {
+                extreme = extremes[j];
             }
         }
-        NEXT
+    } else {
+        WHILE {
+            const npy_DTYPE0* pa = PA(DTYPE0);
+            FOR {
+                const npy_DTYPE0 ai = pa[it.i * it.stride];
+                if (ai COMPARE extreme) {
+                    extreme = ai;
+                    allnan = 0;
+                }
+            }
+            NEXT
+        }
     }
-    if (allnan) extreme = BN_NAN;
+    if (allnan) {
+        extreme = BN_NAN;
+    }
     BN_END_ALLOW_THREADS
     return PyFloat_FromDouble(extreme);
 }
@@ -487,8 +538,9 @@ REDUCE_ONE(NAME, DTYPE0) {
 /* dtype end */
 
 /* dtype = [['int64'], ['int32']] */
+BN_OPT_3
 REDUCE_ALL(NAME, DTYPE0) {
-    npy_DTYPE0 ai, extreme = BIG_INT;
+    npy_DTYPE0 extreme = BIG_INT;
     INIT_ALL
     if (SIZE == 0) {
         VALUE_ERR("numpy.NAME raises on a.size==0 and axis=None; "
@@ -496,12 +548,24 @@ REDUCE_ALL(NAME, DTYPE0) {
         return NULL;
     }
     BN_BEGIN_ALLOW_THREADS
-    WHILE {
-        FOR {
-            ai = AI(DTYPE0);
-            if (ai COMPARE extreme) extreme = ai;
+    if (REDUCE_CONTIGUOUS) {
+        WHILE {
+            const npy_DTYPE0* pa = PA(DTYPE0);
+            FOR {
+                const npy_DTYPE0 ai = pa[it.i];
+                extreme = ai INTCOMP extreme ? ai : extreme;
+            }
+            NEXT
         }
-        NEXT
+    } else {
+        WHILE {
+            const npy_DTYPE0* pa = PA(DTYPE0);
+            FOR {
+                const npy_DTYPE0 ai = pa[it.i * it.stride];
+                extreme = ai INTCOMP extreme ? ai : extreme;
+            }
+            NEXT
+        }
     }
     BN_END_ALLOW_THREADS
     return PyLong_FromLongLong(extreme);