Largest benefits seen by `float32` using `clang` but beneficial across the board.

```
$ asv compare HEAD^ HEAD -s --sort ratio --only-changed
       before           after         ratio
     [eaf2bb00]       [e95171d0]
     <anynan_reduce~1>       <nanmin_reduce>
-        888±20ns         799±30ns     0.90  reduce.Time1DReductions.time_nanmin('int64', (1000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     1.02±0.02μs         882±20ns     0.87  reduce.Time1DReductions.time_nanmin('float64', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-      6.49±0.1ms       5.51±0.2ms     0.85  reduce.Time1DReductions.time_nanmin('float64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        661±20μs         544±30μs     0.82  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        53.8±1μs       43.0±0.9μs     0.80  reduce.Time1DReductions.time_nanmin('int64', (100000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        660±10μs         526±20μs     0.80  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.27±0.03ms         939±30μs     0.74  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.21±0.01μs         892±10ns     0.74  reduce.Time1DReductions.time_nanmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.27±0.04ms         925±40μs     0.73  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.13±0.02ms         741±20μs     0.66  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.13±0.01ms         738±10μs     0.65  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.10±0.01ms         687±10μs     0.63  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        61.1±2μs         38.1±1μs     0.62  reduce.Time1DReductions.time_nanmin('float64', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        947±30ns         586±20ns     0.62  reduce.Time1DReductions.time_nanmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.11±0.01ms          681±9μs     0.62  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        935±20ns         534±30ns     0.57  reduce.Time1DReductions.time_nanmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-      6.05±0.1ms       2.87±0.1ms     0.47  reduce.Time1DReductions.time_nanmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-      5.97±0.1ms      2.76±0.04ms     0.46  reduce.Time1DReductions.time_nanmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        601±10μs          253±6μs     0.42  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        607±10μs          254±5μs     0.42  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-      59.1±0.6μs       23.6±0.4μs     0.40  reduce.Time1DReductions.time_nanmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-         593±7μs          211±6μs     0.36  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        594±10μs          208±3μs     0.35  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     2.62±0.02μs         846±20ns     0.32  reduce.Time1DReductions.time_nanmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     2.61±0.04μs         777±20ns     0.30  reduce.Time1DReductions.time_nanmin('float64', (1000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-      9.07±0.2ms      2.67±0.04ms     0.29  reduce.Time1DReductions.time_nanmin('float32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        58.1±1μs       17.1±0.6μs     0.29  reduce.Time1DReductions.time_nanmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        902±30μs         210±10μs     0.23  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-      23.0±0.1ms      5.28±0.08ms     0.23  reduce.Time1DReductions.time_nanmin('float64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        892±20μs          202±7μs     0.23  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     2.31±0.02ms         493±20μs     0.21  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     2.31±0.01ms          492±6μs     0.21  reduce.Time2DReductions.time_nanmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-        88.2±1μs       16.9±0.7μs     0.19  reduce.Time1DReductions.time_nanmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-       226±0.6μs       32.2±0.6μs     0.14  reduce.Time1DReductions.time_nanmin('float64', (100000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-      22.8±0.1ms      2.59±0.05ms     0.11  reduce.Time1DReductions.time_nanmin('float32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     2.28±0.01ms          178±6μs     0.08  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-     2.27±0.01ms          175±8μs     0.08  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
-       226±0.5μs       14.0±0.4μs     0.06  reduce.Time1DReductions.time_nanmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCclang-CXXclang++]
```